Revert "Fix interpolation/STRFMT overflow issues"
This reverts commit 992be3fd9bd5fd62d2955bb494e14ce5e47c73d4.

+++ include/asm/format.h
@@ -28,9 +28,9 @@ struct FormatSpec {
 	bool prefix;
 	bool alignLeft;
 	bool padZero;
-	size_t width;
+	uint8_t width;
 	bool hasFrac;
-	size_t fracWidth;
+	uint8_t fracWidth;
 	int type;
 	bool valid;
 };
